% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{get_package_functions}
\alias{get_package_functions}
\title{This function uses ellmer to query llama3.2:3b from Ollama to get a list of
exported functions from a specified R package. Results may not be
completely accurate as they are generated by an LLM.}
\usage{
get_package_functions(package_name, model = "llama3.2:3b", temperature = 0.2)
}
\arguments{
\item{package_name}{Character. The name of the R package to query.}

\item{model}{Character. The name of the Ollama model to use. Default is "llama3.2:3b".}

\item{temperature}{Numeric. Temperature setting for the model. Default is 0.2.}
}
\value{
Character vector of function names supposedly exported by the package.
}
\description{
This function uses ellmer to query llama3.2:3b from Ollama to get a list of
exported functions from a specified R package. Results may not be
completely accurate as they are generated by an LLM.
}
\examples{
\dontrun{
get_package_functions("dplyr")
}
}
